+++ packages/react-filerobot-image-editor/src/context/defaultConfig.js
@@ -39,6 +39,8 @@ export default {
   },
   [TOOLS_IDS.IMAGE]: {
     fill: undefined,
+    disableUpload: false,
+    gallery: [],
   },
   [TOOLS_IDS.RECT]: {
     cornerRadius: 0,
@@ -94,6 +96,7 @@ export default {
   closeAfterSave: false,
   defaultSavedImageName: undefined,
   defaultSavedImageType: null,
+  defaultSavedImageQuality: 0.92,
   forceToPngInEllipticalCrop: false,
   useBackendTranslations: true,
   translations: null,